jleague-api/requirements.txt
Flask
flask-sqlalchemy
Anaconda
mysqlclient
pillow